Block swap for GB, JP and MY
