--- VultisigApp/VultisigApp/macOS/Utils/Coin+ChainAction+macOS.swift
@@ -1,16 +1,0 @@
-//
-//  Coin+ChainAction+macOS.swift
-//  VultisigApp
-//
-//  Created by Amol Kumar on 2024-09-27.
-//
-
-#if os(macOS)
-import SwiftUI
-
-extension Array where Element == CoinAction {
-    var filtered: [CoinAction] {
-        return self
-    }
-}
-#endif
